Time: 38 ms (74.66%) | Memory: 29.8 MB (23.10%) - LeetSync

+++ 4323-count-subarrays-with-even-odd-ratio-i/count-subarrays-with-even-odd-ratio-i.cpp
@@ -14,7 +14,7 @@ public:
                 else{
                     y_odd++;
                 }
-                if(y_odd>0 && x_even*b<=a*y_odd){
+                if(y_odd>0 && (long long)x_even*b<=a*(long long)y_odd){
                     valid++;
                 }
             }
